separate folder source created in src for AuthContext.js
